Added signal handlers for stopping/reloading scribe.
When we added ssl support we ran into some issues with puppet trying to communicate with
scribe to get it to apply new config changes etc.  Specifically puppet
didn't know when to use the SSL params for the scribe_ctrl program
and when not to.

Scribe will now listen to SIGTERM and SIGUSR1 to stop and reload the config (respectively).